refactor/thread-a stage 3 — delete config/filters.yml + FilterRegistry (four-file model reached)
