+++ app/views/home/index.html.erb
@@ -46,9 +46,6 @@
   </ul>
 </div>
 
-<%# Canonical filter keys from config/filters.yml (FilterRegistry) — embedded JSON for filter_controller %>
-<%= render "filter_registry_config" %>
-
 <%# Left sidebar navigation %>
 <%= render "sidebar" %>
